api_gateway/requirements.txt
# Database and Migrations
lmdb 
msgpack

# FastAPI and WebSocket
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
python-socketio 
python-engineio

# ADK and AI
google-adk>=0.1.0
google-generativeai>=0.3.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0

# Development
pytest>=8.0.0
pyjwt
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
black>=24.0.0


